% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_in_unit.R
\name{age_in_unit}
\alias{age_in_unit}
\title{Age in some units}
\usage{
age_in_unit(
  DOB,
  REFDATE = Sys.Date(),
  years = TRUE,
  months = TRUE,
  weeks = TRUE,
  days = TRUE
)
}
\arguments{
\item{DOB, REFDATE}{Two dates}

\item{years, months, weeks, days}{the units.}
}
\description{
Age in some units
}
\examples{
\dontshow{if (require("lubridate")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
DOB <- as.Date("1989-08-05")
TODAY <- Sys.Date()
age_in_unit(DOB, TODAY)
\dontshow{\}) # examplesIf}
}
