% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stla.R
\name{bracketifyFile}
\alias{bracketifyFile}
\alias{bracketifySelection}
\title{bracketify}
\usage{
bracketifyFile()

bracketifySelection()
}
\description{
Adpated from \href{https://github.com/stla/bracketify}{stla/bracketify}
}
