% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cor.R
\name{compare_cor}
\alias{compare_cor}
\title{Compare correlations}
\usage{
compare_cor(data, r1, r2, data2 = NULL, by = NULL, ci = 0.95)
}
\arguments{
\item{data}{A data frame}

\item{r1, r2}{Names of variable for the first and second correlations}

\item{data2}{Where to look for the \code{r2} columns (if not provided, looked for
on \code{data}).}

\item{by}{Name of column to split \code{data} by. The column must have only 2
unique values. If provided, the correlation between \code{r1} is compared
between the two groups.}

\item{ci}{Confidence level for correlations.}
}
\value{
A list of two data frames:
\enumerate{
\item The two correlations + their CIs
\item The test results
}
}
\description{
This function is a wrapper around \link[psych:r.test]{psych::r.test}.
}
\details{
\itemize{
\item If \code{data2} is provided, the correlation between \code{r1} (in \code{data}) and \code{r2} (in \code{data2}) is compared.
\itemize{
\item If \code{r2} not provided, the correlation between \code{r1} (in \code{data}) and \code{r1} (in \code{data2}) is compared
}
\item If \code{by} is provided \code{r1} (in \code{data}) is compared between the 2 groups.
\item Else, a test for the difference of two dependent correlations is conducted.
}
}
\examples{
# Test dependent correlations ------------------
## different variables
compare_cor(mtcars, r1 = c("mpg", "hp"), r2 = c("drat", "am"))

## 1 shared variable
compare_cor(mtcars, r1 = c("mpg", "hp"), r2 = c("mpg", "am"))



# Test independent correlations -----------------
## Different data sets
compare_cor(data = mtcars, r1 = c("mpg", "hp"),
            data2 = iris, r2 = c("Sepal.Length", "Sepal.Width"))

## Groups
compare_cor(mtcars, r1 = c("mpg", "hp"), by = "am")


}
