% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_freqs.R
\name{compare_freqs}
\alias{compare_freqs}
\title{Compare the frequencies of levels of a factor}
\usage{
compare_freqs(f, adjust = stats::p.adjust.methods, correct = TRUE)
}
\arguments{
\item{f}{A factor vector.}

\item{adjust}{Method for correcting p-values. See \code{\link[stats:p.adjust]{stats::p.adjust}}.}

\item{correct}{a logical indicating whether to apply continuity correction when computing the test statistic.}
}
\description{
Using \code{\link[stats:mcnemar.test]{stats::mcnemar.test}} for comparing dependent proportions.
}
\examples{
f <- c(
  rep("A", 12),
  rep("B", 45),
  rep("C", 42),
  rep("D", 20)
)

compare_freqs(f)

}
