% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-deprecated.R
\name{crop_coord_polar}
\alias{crop_coord_polar}
\alias{pred_strength}
\title{Deprecated functions}
\usage{
crop_coord_polar(...)

pred_strength(...)
}
\arguments{
\item{...}{Arguments (unused)}
}
\description{
Deprecated functions
}
