% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbind.R
\name{dbind}
\alias{dbind}
\title{Bind matrices diagonally}
\usage{
dbind(..., .fill = 0)
}
\arguments{
\item{...}{Numerical matrices.}

\item{.fill}{Value to fill the off-"diagonal" values.}
}
\description{
Bind matrices diagonally
}
