% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{Transform means a (co)variances using the \emph{delta} method}
\usage{
delta_method(..., .means, .V)
}
\arguments{
\item{...}{Unquoted transformations. See example.}

\item{.means}{A named vector of means.}

\item{.V}{A covariance matrix with named columns and rows corresponding to
\code{names(.means)}.}
}
\value{
A list with \code{means} and \code{V} of the transformed variables. (row/col)
names are of the \code{g} transformations. See examples.
}
\description{
Transform means a (co)variances using the \emph{delta} method
}
\note{
Most of this function is mostly copied from \code{\link[msm:deltamethod]{msm::deltamethod()}}.
}
\examples{

M <- sapply(mtcars, mean)
V <- cov(mtcars)

delta_method(
  (mpg^2)/hp, log1p(am),
  .means = M, .V = V
)

}
