% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ft_ES_CIs.R
\name{F_to_PVE_CI}
\alias{F_to_PVE_CI}
\alias{t_to_PVE_CI}
\alias{t_to_r_CI}
\alias{F_to_r_CI}
\alias{t_to_d_CI}
\alias{F_to_d_CI}
\title{Convert F and t values to Noncentrality-Based Confidence Intervals}
\usage{
F_to_PVE_CI(
  f,
  df,
  df_error,
  conf.level = 0.9,
  type = c("eta2", "epsilon2", "omega2")
)

t_to_PVE_CI(
  t,
  df_error,
  conf.level = 0.9,
  type = c("eta2", "epsilon2", "omega2")
)

t_to_r_CI(t, df_error, conf.level = 0.95)

F_to_r_CI(f, df, df_error, conf.level = 0.95)

t_to_d_CI(t, df_error, conf.level = 0.95, pooled = FALSE)

F_to_d_CI(f, df, df_error, conf.level = 0.95, pooled = FALSE)
}
\arguments{
\item{f, t}{The t or the F statistics.}

\item{df, df_error}{Degrees of freedom of numerator or of the error estimate (i.e., the residuals).}

\item{conf.level}{Confidence level.}

\item{type}{See \link[effectsize:F_to_eta2]{the corresponding functions in effectsize}.}
}
\description{
These methods searches for a the best \code{ncp} (non-central parameters) for either
the t distribution (for r and Cohen's d effect
sizes) of the F distribution (for the PVE effect sizes) for the desired tail-probabilities,
and then convert these \code{ncp}s to the corresponding effect sizes using \code{effectsize}'s
conversion functions.
}
\details{
All effect sizes are converted using simple conversion formulas. See formulas for
\link[effectsize:F_to_eta2]{partial variance explained (PVE)} and for \link[effectsize:t_to_d]{d and r}.

CI for PVE are based on Steiger (2004, eq. 16), and for Cohen's d are based on Cumming and Finch (2001, eq. 17-18).
Read \href{https://tinyurl.com/srqd3ys}{here} for more.
}
\examples{
# CI to eta2
if (requireNamespace("MBESS", quietly = TRUE)) {
  cbind(MSBMisc = F_to_PVE_CI(4, 3, 123)[-2],
        MBESS = unlist(MBESS::ci.pvaf(
          4, 3, 123, N = 127, conf.level = 0.9
        ))[c(1, 3)])
}

# CI for d
if (requireNamespace("MBESS", quietly = TRUE)) {
  cbind(MSBMisc = t_to_d_CI(4, 68)[-2],
        MBESS = unlist(MBESS::ci.smd(4, n.1 = 40, n.2 = 30))[-2])
}

# CI for r
res <- cor.test(iris[[1]], iris[[2]])
cbind(MSBMisc = t_to_r_CI(res$statistic, res$parameter)[-2],
      stats = res$conf.int)
}
\references{
\itemize{
\item Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.
\item Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and calculation of confidence intervals that are based on central and noncentral distributions. Educational and Psychological Measurement, 61(4), 532-574.
}
}
