% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_reliability.R
\name{fa_reliability}
\alias{fa_reliability}
\alias{fa_reliability.fa}
\alias{fa_reliability.lavaan}
\title{Calculate reliability from E/CFA}
\usage{
fa_reliability(x, ...)

\method{fa_reliability}{fa}(x, keys = NULL, threshold = 0, labels = NULL, ...)

\method{fa_reliability}{lavaan}(x, ...)
}
\arguments{
\item{x}{E/CFA model (e.g., the result from \code{psych::fa()} or
\code{lavaan::cfa()}).}

\item{...}{For \code{lavaan} objects, arguments passed to \code{semTools::compRelSEM()}}

\item{keys}{optional, see ?psych::make.keys}

\item{threshold}{which values from the loadings should be used? Only used if
\code{keys = NULL}.}

\item{labels}{optional factor labels}
}
\description{
Calculate reliability from E/CFA
}
\examples{
\dontshow{if (require("psych") && require("GPArotation")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data("Harman74.cor")
EFA <- psych::fa(Harman74.cor$cov, 4)

fa_reliability(EFA)
\dontshow{\}) # examplesIf}
\dontshow{if (require("lavaan") && require("semTools")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

CFA <- lavaan::cfa(HS.model, data = lavaan::HolzingerSwineford1939)

fa_reliability(CFA)

\dontshow{\}) # examplesIf}
}
\author{
\href{https://wiernik.org/}{Brenton M. Wiernik}
}
