% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_grid.R
\name{get_data_for_grid}
\alias{get_data_for_grid}
\alias{get_data_for_grid.data.frame}
\alias{get_data_for_grid.ggeffects}
\alias{get_data_for_grid.emmGrid}
\title{Get raw data for plotting with model predictions}
\usage{
get_data_for_grid(grid, model, residualize = FALSE, collapse_by = FALSE, ...)

\method{get_data_for_grid}{data.frame}(
  grid,
  model,
  residualize = FALSE,
  collapse_by = FALSE,
  pred_name,
  ...
)

\method{get_data_for_grid}{ggeffects}(
  grid,
  model,
  residualize = FALSE,
  collapse_by = FALSE,
  protect_names = TRUE,
  ...
)

\method{get_data_for_grid}{emmGrid}(
  grid,
  model,
  residualize = FALSE,
  collapse_by = FALSE,
  protect_names = TRUE,
  ...
)
}
\arguments{
\item{grid}{A data grid with predictions}

\item{model}{The statistical model}

\item{residualize}{Should data be residualized?}

\item{collapse_by}{Name of grouping variable to collaple across. If \code{TRUE}
name of grouping variable is automatically detected from the model.}

\item{pred_name}{Name of column that has the predictions in the data grid}

\item{protect_names}{Logical, if \code{TRUE}, preserves column names from the
\code{ggeffects} object.}
}
\description{
Get raw data for plotting with model predictions
}
