% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgl2.R
\name{gt}
\alias{gt}
\alias{lt}
\alias{eq}
\alias{neq}
\alias{leq}
\alias{geq}
\alias{\%>>\%}
\alias{\%<<\%}
\alias{\%==\%}
\alias{\%!=\%}
\alias{\%<=\%}
\alias{\%>=\%}
\title{Compare multiple vectors}
\usage{
gt(...)

lt(...)

eq(...)

neq(...)

leq(...)

geq(...)

x \%>>\% y

x \%<<\% y

x \%==\% y

x \%!=\% y

x \%<=\% y

x \%>=\% y
}
\arguments{
\item{x, y, ...}{Vectors, typically numerical, to be compared.}
}
\value{
A logical vector. For the operator, a \code{lat_y} attribute stores the
last RHS values from the comparisons (strip away with \code{as.vector()}). See
examples.
}
\description{
Compare multiple vectors
}
\examples{
x <- c(1, 3, 1, 1, 2)
y <- c(2, 2, 1, 1, 1)
z <- c(3, 1, 1, 2, 1)


lt(x, y, z) # >
eq(x, y, z) # ==
neq(x, y, z) # !=
leq(x, y, z) # <=
geq(x, y, z) # >=

gt(x, y, z) # <

# same as
x \%>>\% y \%>>\% z

# same as
x > y & y > z

# Operators can be mixed!

x \%>>\% y \%==\% z

# Or broken
(l1 <- x \%>>\% y)

(l2 <- l1 \%==\% z)

# same as
x > y & y == z

as.vector(l2)

}
