% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_has_data.R
\name{has_any_data}
\alias{has_any_data}
\alias{has_all_data}
\alias{missing_any_data}
\alias{missing_all_data}
\title{Test rowwise if each row is missing / has all data in select columns}
\usage{
has_any_data(.data, ..., .name)

has_all_data(.data, ..., .name)

missing_any_data(.data, ..., .name)

missing_all_data(.data, ..., .name)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{.name}{Name of the new column with the logical index.}
}
\description{
Test rowwise if each row is missing / has all data in select columns
}
\examples{
data(mtcars)

mtcars[1, 1] <- NA
mtcars[2, ] <- NA

mtcars[1:3, 1:3] |>
  has_any_data(mpg:disp, .name = "has_any") |>
  has_all_data(mpg:disp, .name = "has_all") |>
  missing_any_data(mpg:disp, .name = "missing_any") |>
  missing_all_data(mpg:disp, .name = "missing_all")

}
