% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_matrix.R
\name{index_matrix}
\alias{index_matrix}
\alias{index_matrix.matrix}
\alias{index_matrix<-}
\alias{index_matrix<-.matrix}
\title{Vectorised Extract and Replace from Matrix}
\usage{
index_matrix(x, ...)

\method{index_matrix}{matrix}(x, ...)

index_matrix(x, ...) <- value

\method{index_matrix}{matrix}(x, ...) <- value
}
\arguments{
\item{x}{A matrix}

\item{...}{vectors of indices}

\item{value}{vector of values to replace}
}
\description{
Vectorised Extract and Replace from Matrix
}
\examples{
(M <- matrix(1:12, 3, 4))


index_matrix(M, 1:2, 3:4) # get values at (i = 1, j = 3) and (2, 4)

# same as:
c(M[1, 3], M[2, 4])


# replace values
index_matrix(M, 1:2, 3:4) <- NA

M

}
