% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgl.R
\name{is.TRUE}
\alias{is.TRUE}
\alias{is.FALSE}
\alias{allTRUE}
\alias{anyTRUE}
\title{It's just logical}
\usage{
is.TRUE(x)

is.FALSE(x)

allTRUE(...)

anyTRUE(...)
}
\arguments{
\item{x}{\code{\link[base]{raw}}, \code{\link[base]{logical}} or \sQuote{number-like} vectors (i.e., of
    types \code{\link[base]{double}} (class \code{\link[base]{numeric}}),
    \code{\link[base]{integer}} and \code{\link[base]{complex}}), or objects for
    which methods have been written.}

\item{...}{Values to be tested.}
}
\description{
It's just logical
}
\examples{
x <- list(1, TRUE, list(TRUE), FALSE, "Hello world!")

is.TRUE(x)

is.FALSE(x)

allTRUE(TRUE, FALSE, stop("NOOOO"))

anyTRUE(FALSE, TRUE, stop("NOOOO"))

}
