% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_converter.R
\name{lnorm_mean}
\alias{lnorm_mean}
\alias{lnorm_median}
\alias{lnorm_var}
\alias{lnorm_sd}
\title{Functions to convert parameters of a log-normal distribution to meaningfull
values on the response scale.}
\usage{
lnorm_mean(meanlog, sdlog, ...)

lnorm_median(meanlog, ...)

lnorm_var(meanlog, sdlog, ...)

lnorm_sd(meanlog, sdlog, ...)
}
\arguments{
\item{meanlog, sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{...}{Not used}
}
\description{
Functions to convert parameters of a log-normal distribution to meaningfull
values on the response scale.
}
\examples{
x <- rlnorm(1e4, meanlog = 1.5, sdlog = 1.2)

m <- lm(log(x) ~ 1)

meanlog <- coef(m)
sdlog <- sigma(m)

lnorm_mean(meanlog, sdlog)
mean(x)

lnorm_median(meanlog, sdlog)
median(x)

lnorm_sd(meanlog, sdlog)
sd(x)


}
