% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_lnorm.R
\name{logLik_lnorm}
\alias{logLik_lnorm}
\alias{AIC_lnorm}
\alias{BIC_lnorm}
\title{Information criteria for log-normal models}
\usage{
logLik_lnorm(object, REML = FALSE)

AIC_lnorm(object, k = 2, REML = FALSE)

BIC_lnorm(object, REML = FALSE)
}
\arguments{
\item{object}{A fitted model object. The model must meet all of the
following (will throw an error if not met):
\enumerate{
\item A Gaussian likelihood with an identity link function.
\item The LHS of the model's formula must use the \code{log()} function.
\item No weights (not yet supported).
}}

\item{REML}{Only \code{FALSE} supported.}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\description{
Log-likelihood (and by extension \emph{AIC} and \emph{BIC}) for log-normal models fit
with \code{stats::lm(log(y) ~ ...)} are computed with \code{stats::dnorm(log(y), ...)}
instead of with \code{stats::dlnorm(y, ...)}, which makes comparing different
families difficult. This function is aimed at rectifying this. See examples.
}
\note{
\code{REML} is not (yet) supported. Make sure you are comparing correct
LL/AIC/BIC values.
}
\examples{

data("mtcars")
mtcars$mpg <- floor(mtcars$mpg)

model_lnorm <- lm(log(mpg) ~ factor(cyl), mtcars)
model_norm <- lm(mpg ~ factor(cyl), mtcars)
model_pois <- glm(mpg ~ factor(cyl), mtcars, family = poisson())

# No, that first one is wrong...
(aics <- AIC(model_lnorm, model_norm, model_pois))

aics[1, "AIC"] <- AIC_lnorm(model_lnorm)
aics # better!

\dontshow{if (require("lme4")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Should support any model really... =====================
model_lnorm <- lme4::lmer(log(mpg) ~ factor(cyl) + (1 | gear), mtcars, REML = FALSE)
model_norm <- lme4::lmer(mpg ~ factor(cyl) + (1 | gear), mtcars, REML = FALSE)
model_pois <- lme4::glmer(mpg ~ factor(cyl) + (1 | gear), mtcars, family = poisson())

# No, that first one is wrong...
(aics <- AIC(model_lnorm, model_norm, model_pois))

aics[1, "AIC"] <- AIC_lnorm(model_lnorm)
aics # better!
\dontshow{\}) # examplesIf}
}
