% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-ML.R
\name{metric_by_event}
\alias{metric_by_event}
\title{Compute class-wise metrics for a metric set in a multiclass setting}
\usage{
metric_by_event(data, metric_set, truth, estimate, ..., .all = FALSE)
}
\arguments{
\item{data}{A data frame containing the columns specified in \code{truth}, \code{estimate}, and \code{...}.}

\item{metric_set}{A metric set function (e.g., \code{yardstick::metric_set()}) that includes class-wise metrics.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{factor} vector.}

\item{...}{Not currently used.}

\item{.all}{A logical value indicating whether to include all metrics from the metric set (including those that do not use macro/micro estimators).}
}
\value{
A tibble with the same columns to be expected from the metric set,
plus the following additional columns: \code{.class} indicating the class for which
the metric was computed. For these rows, \code{.estimator} is set to \code{"event"} to
indicate that these are event-level metrics.
}
\description{
This function computes class-wise metrics for a metric set in a multiclass
setting by binarizing the outcome variable for each class and applying the
metric set to each binary version of the data.
}
\seealso{
Other ML4Psych: 
\code{\link{jaccard_avg}()},
\code{\link{pred_strength}()}
}
\concept{ML4Psych}
