% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stla.R
\name{Paste}
\alias{Paste}
\alias{PasteRoxygen}
\alias{Copy}
\alias{CopyRoxygen}
\title{Copy / Paste as Comment / Roxygen}
\usage{
Paste(prefix = "#> ")

PasteRoxygen()

Copy(prefix = "#> ")

CopyRoxygen()
}
\arguments{
\item{prefix}{The prefix to use for the copied text.}
}
\description{
Adpated from \href{https://github.com/stla/pasteAsComment}{stla/pasteAsComment}
}
