% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHP.R
\name{php.t}
\alias{php.t}
\alias{php.F}
\alias{php.z}
\alias{php.chisq}
\alias{php.guf}
\title{Function for Post-Hoc Power analysis}
\usage{
php.t(tval, p, df = Inf, alpha = 0.05)

php.F(Fval, p, df1, df2 = Inf, alpha = 0.05)

php.z(zval, p, alpha = 0.05)

php.chisq(chisqval, p, df, alpha = 0.05)

php.guf(p, alpha = 0.05)
}
\arguments{
\item{tval, zval, Fval, chisqval}{Observed test statistic}

\item{p}{Observed p-value (used if test statistic not supplied)}

\item{df, df1, df2}{Test statistics' degrees of freedom}

\item{alpha}{Confidence level of the test.}
}
\description{
Based on \href{https://stat.uiowa.edu/sites/stat.uiowa.edu/files/techrep/tr378.pdf}{Lenth (2007) \emph{Post Hoc Power: Tables and Commentary}}.
\code{php.guf()} give's Lenth's \emph{grand unified formula for post hoc power}.
}
