% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_worm_plot.R
\name{qq_worm_plot}
\alias{qq_worm_plot}
\title{Create a worm plot}
\usage{
qq_worm_plot(x, distribution = "norm", return = c("plot", "data"), ...)
}
\arguments{
\item{x}{A numerical vector}

\item{distribution}{Name of a distribution, matching the \verb{d*}, \verb{p*} and \verb{q*} function names.}

\item{return}{a \code{ggplot} or a data frame?}

\item{...}{Args assed to \verb{d*}, \verb{p*} and \verb{q*} functions.}
}
\description{
Create a worm plot
}
\details{
\href{https://twitter.com/mattansb/status/1199936633413476358}{Some related tweets}.
}
\examples{
x <- rnorm(100)
qq_worm_plot(x)
qq_worm_plot(x, return = "data")

x <- rbeta(100, shape1 = 2, shape2 = 3)
qq_worm_plot(x, distribution = "beta", shape1 = 2, shape2 = 3)

\dontrun{
  x <- rexp(100)
  qq_worm_plot(x, distribution = "exp")

  x <- rpois(100, lambda = 15)
  qq_worm_plot(x, distribution = "pois", lambda = 15)

  x <- rt(100, df = 3)
  qq_worm_plot(x, distribution = "t", df = 3)

  x <- runif(100)
  qq_worm_plot(x, distribution = "unif")
}

}
