% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{\eqn{R^2} Overkill}
\usage{
R2(pred, obs, type = 1, na.rm = TRUE)
}
\arguments{
\item{pred}{The predicted values by some model; typically the result of a
call to \code{\link[=predict]{predict()}}.}

\item{obs}{The true observed values.}

\item{type}{Which of the 8 \enq{R^2} to use. See details.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\description{
Way too many ways to calculate \eqn{R^2}.
}
\details{
The types of \eqn{R^2}:
\itemize{
\item \eqn{R^2_1 = 1 - \sum (y-\hat{y})^2 / \sum (y-\bar{y})^2}
\item \eqn{R^2_2 = \sum (\hat{y}-\bar{y})^2 / \sum (y-\bar{y})^2}
\item \eqn{R^2_3 = \sum (\hat{y}-\bar{\hat{y}})^2 / \sum (y-\bar{y})^2}
\item \eqn{R^2_4 = 1 - \sum (e-\bar{e})^2 / \sum (y-\bar{y})^2}
\item \eqn{R^2_5 = } squared multiple correlation coefficient between the regressand and the regressors
\item \eqn{R^2_6 = r_{y,\hat{y}}^2}
\item \eqn{R^2_7 = 1 - \sum (y-\hat{y})^2 / \sum y^2}
\item \eqn{R^2_8 = \sum \hat{y}^2 / \sum y^2}
}
}
\examples{
X <-  c(1, 2, 3, 4, 5, 6)
Y <- c(15, 37, 52, 59, 83, 92)

m1 <- lm(Y ~ X)
m2 <- lm(Y ~ 0 + X)
m3 <- lm(log(Y) ~ X)

# Table 2
data.frame(
  mod1 = sapply(1:8, R2, pred = predict(m1), obs = Y),
  mod2 = sapply(1:8, R2, pred = predict(m2), obs = Y),
  mod3 = sapply(1:8, R2, pred = exp(predict(m3)), obs = Y)
)

}
\references{
Kvålseth, T. O. (1985). Cautionary note about R 2. The American Statistician, 39(4), 279-285.
}
