% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-HLM.R
\name{r2_pseudo}
\alias{r2_pseudo}
\alias{V_table}
\title{Compute pseudo R-square for mixed models}
\usage{
r2_pseudo(mf, mr, mnull = mr)

V_table(model)
}
\arguments{
\item{mf}{The full model}

\item{mr}{The reduced model}

\item{mnull}{The empty model, where some variance of interest is unmodeled
(an "unconditional" model)}

\item{model}{A fitted model object}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{grp}: The name of the grouping factor
\item \code{var}: The name of the variance component
\item \code{r2}: The pseudo R-square for that variance component
}
}
\description{
This function computes the pseudo R-square for mixed models by quantifying
the reduction on the variance-covariance between a full model and a reduced
model.
}
\details{
Variance components are extracted from the 3 models, and where they match
(groups and terms must match \emph{exactly}), we compute:

\deqn{R^2 = \frac{V_{full} - V_{restricted}}{V_{null}}}

This gives the increase in variance explained by the full model compared to
the reduced model. When the restricted model is the empty model, this is the
same as:

\deqn{R^2 = 1 - \frac{V_{full}}{V_{restricted}}}
\subsection{Supported model classes}{
\itemize{
\item \code{{stats}}: \code{lm}, \code{glm}
\item \code{{lme4}}: \code{lmerMod} and \code{glmerMod}
\item \code{{rstanarm}}: \code{stanreg}
\item \code{{brms}}: \code{brmsfit}
\item \code{{glmmTMB}}: \code{glmmTMB}
}
}
}
\examples{
\dontshow{if (require("lme4")) withAutoprint(\{ # examplesIf}

mod0 <- lme4::lmer(Reaction ~ 1 + (1 | Subject), data = lme4::sleepstudy)
mod1 <- lme4::lmer(Reaction ~ Days + (1 | Subject), data = lme4::sleepstudy)
mod2 <- lme4::lmer(Reaction ~ Days + (Days | Subject), data = lme4::sleepstudy)

r2_pseudo(mod1, mod0) # Residual variance explained by the fixed effect
r2_pseudo(mod2, mod1, mod0) # Residual variance explained by the random effects
r2_pseudo(mod2, mod0) # Residual variance explained by the mixed effects
\dontshow{\}) # examplesIf}
}
\concept{HLM4Psych}
