% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_SB.R
\name{r_SB}
\alias{r_SB}
\title{Spearman-Brown Split half reliability}
\usage{
r_SB(x, y = NULL, var.equal = TRUE)
}
\arguments{
\item{x}{A correlation or numeric vector}

\item{y}{A numeric vector (ignored if \code{x} is not a vector)}

\item{var.equal}{Assume equal var of \code{x} and \code{y}? (ignored if \code{x} is not a vector)}
}
\description{
Spearman-Brown Split half reliability
}
\examples{
r_SB(1:30,-exp(1/1:30), var.equal = TRUE)

r_SB(1:30,-exp(1/1:30), var.equal = FALSE)

r_SB(0.57)
}
