% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths_layout_maker.R
\name{semPaths_layout_maker}
\alias{semPaths_layout_maker}
\title{Shiny app for manual \code{semPaths} layout}
\usage{
semPaths_layout_maker(object, snap_to_grid = TRUE)
}
\arguments{
\item{object}{Object to get the node names from. \code{\link[semPlot]{semPlotModel}}}

\item{snap_to_grid}{Should nodes be auto snapped to the grid?}
}
\value{
A layout matrix that can be passed to \code{\link[semPlot]{semPaths}} via \code{layout=}.
}
\description{
Shiny app for manual \code{semPaths} layout
}
\examples{
\dontrun{
  library(lavaan)
  library(semPlot)

  mediation_model <- "
    Sepal.Length ~ b * Sepal.Width + c * Petal.Length
     Sepal.Width ~ a * Petal.Length
"
  fit <- sem(mediation_model, data = iris)

  semPaths(fit)


  # With user layout
  m <- semPaths_layout_maker(fit, snap_to_grid = TRUE)

  semPaths(fit, layout = m)
}
}
