% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{seq_range}
\alias{seq_range}
\alias{seq_quantile}
\alias{seq_IQR}
\alias{mean_sd}
\alias{median_mad}
\title{Sequence Generation Based on the Values of a Vector}
\usage{
seq_range(
  x,
  length.out = NULL,
  by = NULL,
  along.with = NULL,
  na.rm = TRUE,
  padding = 0.05
)

seq_quantile(
  x,
  probs,
  length.out = NULL,
  by = NULL,
  along.with = NULL,
  na.rm = TRUE
)

seq_IQR(x, length.out = NULL, by = NULL, along.with = NULL, na.rm = TRUE)

mean_sd(x, na.rm = TRUE, out = c("vector", "data.frame"))

median_mad(x, na.rm = TRUE, out = c("vector", "data.frame"))
}
\arguments{
\item{x}{A numeric vector}

\item{length.out}{desired length of the sequence. If no other arguments are
valued, defaults to 20.}

\item{by}{number: increment of the sequence.}

\item{along.with}{take the length from the length of this argument.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}

\item{padding}{Padding factor for the range.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{out}{If \code{"data.frame"} can be used as a summary function in \code{ggplot2}.}
}
\description{
Sequence Generation Based on the Values of a Vector
}
\examples{
set.seed(1)
x <- rt(100, df = 3)
seq_range(x, length.out = 5)
seq_IQR(x, length.out = 5)
seq_quantile(x, c(.05, .95), length.out = 5)

mean_sd(x)

\dontshow{if (require("ggplot2")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
ggplot(mtcars, aes(cyl, mpg)) +
  stat_summary(aes(color = "Mean (SD)"),
    fun.data = mean_sd,
    fun.args = list(out = "data.frame")
  ) +
  stat_summary(aes(color = "Median (MAD)"),
    fun.data = median_mad,
    fun.args = list(out = "data.frame"),
    position = position_nudge(x = 0.5)
  )
\dontshow{\}) # examplesIf}
}
