% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_effects.R
\name{simple_effects}
\alias{simple_effects}
\title{Compute Simple Effects Omnibus Tests}
\usage{
simple_effects(model, effect, inside, ...)
}
\arguments{
\item{model}{The model.}

\item{effect}{The name of the required simple effect. e.g., \code{"A"} for a
simple effect of \emph{A}, or \code{"A:B"} for a simple \emph{A-by-B} interaction.}

\item{inside}{A vector of the name(s) of the variable(s) within whose levels
the \code{effect} will be tested. Can also be the name of an interaction (e.g.,
\code{"B:C"}). If not specified, will use all the terms not in \code{effect}.}

\item{...}{Passed to \code{emmeans::joint_tests()}, e.g., \code{cov.reduce}, \code{at}, etc.}
}
\description{
This is a wrapper for \code{emmeans::joint_tests()} that provides an easy way to
specify which simple effects we wish to test, and within what variable(s).
}
\examples{
\dontshow{if (require("ggplot2") && require("insight") && require("stringr")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(afex)

data(obk.long, package = "afex")
A <- aov_car(value ~ treatment * gender + Error(id/(phase*hour)),
             data = obk.long)

simple_effects(A, effect = "treatment")

simple_effects(A, effect = "treatment:phase")

simple_effects(A, effect = "phase", inside = "treatment")

simple_effects(A, effect = "phase", inside = c("treatment", "gender"))
# simple_effects(A, effect = "phase", inside = "treatment:gender") # same

simple_effects(A, effect = "phase", inside = c("treatment", "gender"),
               at = list(gender = "F"))

simple_effects(A, effect = "phase:treatment", inside = "gender")

\dontshow{\}) # examplesIf}
}
