% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_demo_apps.R
\name{stat_demo_apps}
\alias{stat_demo_apps}
\title{Some stats demos}
\usage{
stat_demo_apps(demo = c("paired ttest", "truncated correlation"))
}
\arguments{
\item{demo}{Which demo to show?}
}
\description{
Some stats demos
}
\details{
\subsection{paired ttest}{
Shows a paired vs un-paired ttest, and how these differences are
affected by the correlation.
}

\subsection{truncated correlation}{
Demo of how truncation affects correlations and doesn't affect MSE.
}
}
