% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_weights.R
\name{contrast_weights}
\alias{contrast_weights}
\alias{cw}
\title{Build Contrast Weights}
\usage{
contrast_weights(...)

cw(...)
}
\arguments{
\item{...}{Can be:
\itemize{
\item Unnamed scalars.
\item (Possibly named) vectors if equal length
}}
}
\value{
Depending on input, either a vector or a data frame of scaled weights.
}
\description{
To be used ideally with \link[emmeans:contrast]{emmeans::contrast}. Each contrasts is tested (sum to
0?) and scaled so that all positive weights sum to 1 (and all negative
weights to -1).
}
\examples{
library(emmeans)

mod <- lm(mpg ~ factor(cyl) * am, mtcars)

(emms <- emmeans(mod, ~ cyl + am))

my_contrasts <- data.frame("squares" = c(-1, 2, -1),
                           "4 vs 6" = c(-30, 30, 0),
                           check.names = FALSE)

(my_contrasts2 <- cw(my_contrasts))

contrast(emms, method = my_contrasts, by = "am")
contrast(emms, method = my_contrasts2, by = "am") # estimate is affected!

contrast(emms, interaction = list(cyl = my_contrasts2, am = "pairwise"))

}
