% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_coord_polar.R
\name{crop_coord_polar}
\alias{crop_coord_polar}
\title{Crop \code{coord_polar}}
\usage{
crop_coord_polar(
  plot,
  start = 0,
  end = 2 * pi,
  padding = 0.02,
  fix_aspect.ratio = TRUE
)
}
\arguments{
\item{plot}{A \code{ggplot}}

\item{start, end}{The angular values (in radians) marking where the plot
starts and ends.}

\item{padding}{How much padding to add around the crop?}

\item{fix_aspect.ratio}{Should the aspect ratio be fixed?}
}
\description{
Crop \code{coord_polar}
}
\examples{
\dontshow{if (require("ggplot2") && require("ggtrace")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(ggplot2)

polar_plot <- ggplot(mtcars ,aes(hp, mpg)) +
  geom_point() +
  geom_smooth(method = "lm") +
  expand_limits(y = c(0, 60)) +
  coord_polar(start = 0, theta = "y")

crop_coord_polar(polar_plot, end = pi)
crop_coord_polar(polar_plot, end = pi/2)
crop_coord_polar(polar_plot, start = 3 * pi / 2, end = pi / 2)



# Also works with facets!
polar_plot_fact <- polar_plot +
  facet_grid(~ gear)

crop_coord_polar(polar_plot_fact, end = pi)

# Use multiple values - one for each facet:
crop_coord_polar(polar_plot_fact,
                 start = c(0, pi/2, pi), end = c(pi/2, pi, 2*pi))
\dontshow{\}) # examplesIf}
}
