% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_coord_polar.R
\name{crop_coord_polar}
\alias{crop_coord_polar}
\title{Crop \code{coord_polar}}
\usage{
crop_coord_polar(
  plot,
  start = 0,
  end = 2 * pi,
  padding = 0.02,
  fix_aspect.ratio = TRUE
)
}
\arguments{
\item{plot}{A \code{ggplot}}

\item{start, end}{The angular values (in radians) marking where the plot
starts and ends.}

\item{padding}{How much padding to add around the crop?}

\item{fix_aspect.ratio}{Should the aspect ratio be fixed?}
}
\description{
Crop \code{coord_polar}
}
\details{
This function uses
\href{https://yjunechoe.github.io/ggtrace/}{\code{{ggtrace}}} by June Choe, adapted
from \href{https://twitter.com/yjunechoe/status/1506672410208051201}{this twitter thread}.
}
\examples{
\dontshow{if (require("ggplot2") && require("ggtrace")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(ggplot2)

polar_plot <- ggplot(mtcars ,aes(hp, mpg)) +
  geom_point() +
  geom_smooth(method = "lm") +
  expand_limits(y = c(0, 60)) +
  coord_polar(start = 0, theta = "y")

crop_coord_polar(polar_plot, end = pi)
crop_coord_polar(polar_plot, end = pi/2)
crop_coord_polar(polar_plot, start = 3 * pi / 2, end = pi / 2)



# Also works with facets!
d <- data.frame(
  x = seq(1, 7, length = 6*5),
  y = rnorm(6*5),
  g = rep(letters[1:6], each = 5)
)

polar_plot_facet <- ggplot(d, aes(x, y)) +
  geom_point(aes(color = x), size = 2) +
  facet_wrap(~g) +
  scale_x_continuous(breaks = seq(0, 6), minor_breaks = NULL) +
  coord_polar()

crop_coord_polar(polar_plot_facet, start = pi)

# Use multiple values - one for each facet:
start <- seq(0, 5) * 2 * pi / 6
end <- start + start[2]

crop_coord_polar(polar_plot_facet,
                 start = start, end = end)


\dontshow{\}) # examplesIf}
}
