% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbind.R
\name{dbind}
\alias{dbind}
\title{Bind matrices diagonally}
\usage{
dbind(..., .fill = NULL)
}
\arguments{
\item{...}{Matrices.}

\item{.fill}{Value to fill the off-"diagonal" values. If \code{NULL}, the value is
the default value of the inputs' mode.}
}
\description{
Bind matrices diagonally
}
\examples{

M1 <- matrix(1:8, 2, 4)
M2 <- matrix(9:14, 2, 3)
dbind(M1, M2)
dbind(M1, M2, .fill = NA)


M1 <- matrix(letters[1:4], 2, 2)
M2 <- matrix(LETTERS[5:10], 3, 2)
dbind(M1, M2)
dbind(M1, M2, .fill = "Banana")


M1 <- matrix(TRUE, 2, 3)
M2 <- matrix(NA, 3, 4)
dbind(M1, M2)


}
