% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{Transform means a (co)variances using the \emph{delta} method}
\usage{
delta_method(..., .means, .V)
}
\arguments{
\item{...}{Unquoted transformations. See example.}

\item{.means}{A named vector of means.}

\item{.V}{A covariance matrix.}
}
\value{
A list with \code{means} and \code{V} of the transformed variables. (row/col)
names are of the \code{g} transformations. See examples.
}
\description{
Transform means a (co)variances using the \emph{delta} method
}
\note{
Most of this function is mostly copied from \code{\link[msm:deltamethod]{msm::deltamethod()}}.
}
\examples{

M <- sapply(mtcars, mean)
V <- cov(mtcars)

delta_method(
  (mpg^2)/hp, log1p(am),
  .means = M, .V = V
)

# Sobel Test ----

data("mtcars")
mod.y <- lm(mpg ~ hp + cyl, mtcars[1:5,])
mod.m <- lm(hp ~ cyl, mtcars[1:5,])

bhat <- c(coef(mod.y), coef(mod.m))[c(2,5)]
Vhat <- dbind(vcov(mod.y), vcov(mod.m))[c(2,5), c(2,5)]

res <- delta_method(
  hp * cyl,
  .means = bhat, .V = Vhat
)

res$means / sqrt(c(res$V))

# Compare:
(bhat[1] * bhat[2]) /
  sqrt(bhat[1]^2 * Vhat[2,2] + bhat[2]^2 * Vhat[1,1])


}
