% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{Transform means a (co)variances using the \emph{delta} method}
\usage{
delta_method(..., .means, .V, return = c("means", "cov", "stddev", "cor"))
}
\arguments{
\item{...}{Unquoted transformations. See example.}

\item{.means}{A named vector of means.}

\item{.V}{A covariance matrix.}

\item{return}{What should be returned?}
}
\value{
A list with one of (optionally named):
\itemize{
\item \code{means} of the transformed variables.
\item \code{cov} (co) variance matrix of the the transformed variables.
\item \code{stddev} standard deviations of the transformed variables (\code{sqrt(diag(cov))}).
\item \code{cor} correlation matrix of the transformed variables. (\code{cov2cor(cov)})
}
}
\description{
Transform means a (co)variances using the \emph{delta} method
}
\note{
Most of this function is mostly copied from \code{msm::deltamethod()}.
}
\examples{

M <- sapply(mtcars, mean)
V <- cov(mtcars)

delta_method(
  (mpg^2) / hp,
  log_am = log1p(am),
  .means = M, .V = V,
  return = "cor"
)

# Sobel Test ----

data("mtcars")
mod.y <- lm(mpg ~ hp + cyl, mtcars[1:5, ])
mod.m <- lm(hp ~ cyl, mtcars[1:5, ])

bhat <- c(coef(mod.y), coef(mod.m))[c(2, 5)]
Vhat <- dbind(vcov(mod.y), vcov(mod.m))[c(2, 5), c(2, 5)]

res <- delta_method(
  hp * cyl,
  .means = bhat, .V = Vhat,
  return = c("means", "stddev")
)

res$means / res$stddev

# Compare:
(bhat[1] * bhat[2]) /
  sqrt(bhat[1]^2 * Vhat[2, 2] + bhat[2]^2 * Vhat[1, 1])

# Special character will give you a bad time...
m <- lm(mpg ~ factor(cyl), mtcars[1:5, ])

bhat <- coef(m)
names(bhat) <- c("cyl4", "cyl6", "cyl8")
V <- vcov(m)

delta_method(cyl4, cyl4 + cyl6, cyl4 + cyl8,
  .means = bhat,
  .V = V
)

}
