% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_lnorm.R
\name{logLik_lnorm}
\alias{logLik_lnorm}
\alias{AIC_lnorm}
\alias{BIC_lnorm}
\title{Information criteria for log-normal models}
\usage{
logLik_lnorm(object)

AIC_lnorm(object, ..., k = 2)

BIC_lnorm(object, ...)
}
\arguments{
\item{object}{a fitted model object for which there exists a
    \code{logLik} method to extract the corresponding log-likelihood, or
    an object inheriting from class \code{logLik}.}

\item{...}{optionally more fitted model objects.}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\description{
Log-likelihood (and by extension \emph{AIC} and \emph{BIC}) for log-normal models fit
with \code{stats::lm(log(y) ~ ...)} are computed with \code{dnorm(log(y), ...)} instead
of with \code{dlnorm(y, ...)}, which makes comparing different families difficult.
This function is aimed at rectifying this. See examples.
}
\examples{

data("mtcars")
mtcars$mpg <- floor(mtcars$mpg)

model_lnorm <- lm(log(mpg) ~ factor(cyl), mtcars)
model_norm <- lm(mpg ~ factor(cyl), mtcars)
model_pois <- glm(mpg ~ factor(cyl), mtcars, family = poisson())

# No, that first one is wrong...
(aics <- AIC(model_lnorm, model_norm, model_pois))

aics[1, "AIC"] <- AIC_lnorm(model_lnorm)
aics # better!


\dontshow{if (require("lme4")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Should support any model really... =====================
model_lnorm <- lme4::lmer(log(mpg) ~ factor(cyl) + (1 | gear), mtcars)
model_norm <- lme4::lmer(mpg ~ factor(cyl) + (1 | gear), mtcars)
model_pois <- lme4::glmer(mpg ~ factor(cyl) + (1 | gear), mtcars, family = poisson())

# No, that first one is wrong...
(aics <- AIC(model_lnorm, model_norm, model_pois))

aics[1, "AIC"] <- AIC_lnorm(model_lnorm)
aics # better!

\dontshow{\}) # examplesIf}
}
