% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHP.R
\name{php.t}
\alias{php.t}
\alias{php.F}
\alias{php.z}
\alias{php.chisq}
\alias{php.guf}
\title{Function for Post-Hoc Power analysis}
\usage{
php.t(tval, p, df = Inf, alpha = 0.05, one.tailed = FALSE)

php.F(Fval, p, df1, df2 = Inf, alpha = 0.05)

php.z(zval, p, alpha = 0.05, one.tailed = FALSE)

php.chisq(chisqval, p, df, alpha = 0.05)

php.guf(p, alpha = 0.05)
}
\arguments{
\item{tval, zval, Fval, chisqval}{Observed test statistic}

\item{p}{Observed p-value (used if test statistic not supplied)}

\item{df, df1, df2}{Test statistics' degrees of freedom}

\item{alpha}{Confidence level of the test.}

\item{one.tailed}{Is the p-value from a one-tailed test?}
}
\description{
Based on \href{https://stat.uiowa.edu/sites/stat.uiowa.edu/files/techrep/tr378.pdf}{Lenth (2007) \emph{Post Hoc Power: Tables and Commentary}}.
\code{php.guf()} give's Lenth's \emph{grand unified formula for post hoc power}.
}
\examples{

lm(hp ~ am, mtcars) |> summary()

php.t(p = 0.18, df = 30)

php.guf(p = 0.18)



# Table 1 - one tailed
expand.grid(
  p = c(.001, .01, .05, .1, .25, .5, .75),
  df = c(1, 2, 5, 10, 20, 50, 200, 1000, Inf)
) |>
  transform(PHP = php.t(p = p, df = df, one.tailed = TRUE)) |>
  stats::reshape(
    direction = "wide",
    idvar = "df", timevar = "p"
  )


# Table 1 - two tailed
expand.grid(
  p = c(.001, .01, .05, .1, .25, .5, .75),
  df = c(1, 2, 5, 10, 20, 50, 200, 1000, Inf)
) |>
  transform(PHP = php.t(p = p, df = df)) |>
  stats::reshape(
    direction = "wide",
    idvar = "df", timevar = "p"
  )


# Table 2
expand.grid(
  p = c(.001, .01, .05, .1, .25, .5, .75),
  df2 = c(1, 2, 5, 10, 20, 50, 200, 1000, Inf),
  df1 = c(2, 3, 4, 10)
) |>
  transform(PHP = php.F(p = p, df1 = df1, df2 = df2)) |>
  stats::reshape(
    direction = "wide",
    idvar = c("df1", "df2"), timevar = "p"
  )

}
