% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{seq_range}
\alias{seq_range}
\alias{mean_sd}
\title{Sequence Generation Baed on Range}
\usage{
seq_range(x, length.out = 20, na.rm = TRUE)

mean_sd(x, na.rm = TRUE, out = c("vector", "data.frame"))
}
\arguments{
\item{x}{A numeric vector}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}

\item{out}{If \code{"data.frame"} can be used as a summary function in \code{ggplot2}.}
}
\description{
Sequence Generation Baed on Range
}
\examples{
set.seed(1)
x <- rnorm(100)
seq_range(x, length.out = 5)

mean_sd(x)

library(ggplot2)
ggplot(mtcars, aes(cyl, mpg)) +
  stat_summary(fun.data = mean_sd, fun.args = list(out = "data.frame"))

}
