% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{seq_range}
\alias{seq_range}
\alias{seq_IQR}
\alias{mean_sd}
\title{Sequence Generation Based on the Values of a Vector}
\usage{
seq_range(x, length.out = NULL, by = NULL, along.with = NULL, na.rm = TRUE)

seq_IQR(x, length.out = NULL, by = NULL, along.with = NULL, na.rm = TRUE)

mean_sd(x, na.rm = TRUE, out = c("vector", "data.frame"))
}
\arguments{
\item{x}{A numeric vector}

\item{length.out}{desired length of the sequence. If no other arguments are
valued, defaults to 20.}

\item{by}{number: increment of the sequence.}

\item{along.with}{take the length from the length of this argument.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{out}{If \code{"data.frame"} can be used as a summary function in \code{ggplot2}.}
}
\description{
Sequence Generation Based on the Values of a Vector
}
\examples{
set.seed(1)
x <- rt(100, df = 3)
seq_range(x, length.out = 5)
seq_IQR(x, length.out = 5)

mean_sd(x)

\dontrun{
library(ggplot2)
ggplot(mtcars, aes(cyl, mpg)) +
  stat_summary(fun.data = mean_sd, fun.args = list(out = "data.frame"))
}
}
