% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_demo_apps.R
\name{stat_demo_apps}
\alias{stat_demo_apps}
\title{Some stats demos}
\usage{
stat_demo_apps(
  demo = c("paired ttest", "truncated correlation", "berksons paradox")
)
}
\arguments{
\item{demo}{Which demo to show? (Partial matching supported)}
}
\description{
Some stats demos
}
\details{
\subsection{paired ttest}{
Shows a paired vs un-paired ttest, and how these differences are
affected by the correlation.
}

\subsection{truncated correlation}{
Demo of how truncation affects correlations and doesn't affect MSE.
}

\subsection{berksons paradox}{
Demo of how How sampling bias can affect estimated correlations and effects.
See related \href{https://www.youtube.com/watch?v=FUD8h9JpEVQ}{Numberphile video}.
}
}
